requirements.txt
gunicorn~=20.1.0
django-environ~=0.8.1
django==4
djangorestframework==3.13.1
drf-yasg==1.20.0
requests~=2.27.1
psycopg2~=2.9.3
django-health-check~=3.17